+++ pytorch_lightning/plugins/training_type/training_type_plugin.py
@@ -311,6 +311,6 @@ class TrainingTypePlugin(Plugin, ABC):
         pass
 
     @property
-    def should_save_checkpoint(self) -> bool:
+    def should_rank_save_checkpoint(self) -> bool:
         """Returns whether the checkpoint should be saved (rank based)"""
         return self.is_global_zero